\name{line}
\docType{data}
\alias{LINE}

\title{Linear regression example}
\description{
The LINE model is a trivial linear regression model with only 5 observations.
It's main use is to allow automated checks of the \pkg{rjags} package.
}
\format{
  A \code{jags.model} object, which must be recompiled before
  use.
}
\keyword{datasets}

