\name{jags.module}
\alias{load.module}
\alias{load.pkg.module}
\alias{unload.module}
\alias{list.modules}
\title{Dynamically load JAGS modules}
\description{
  A JAGS module is a dynamically loaded library that extends the
  functionality of JAGS. These functions load and unload JAGS modules
  and show the names of the currently loaded modules.
  The load.pkg.module function is provided specifically for use
  from other R packages containing JAGS modules and checks for binary
  compatibility between rjags and the indicated package (given by the
  name argument) before loading the module embedded within the package.
  Note that this would normally be done in the .onLoad function.
}
\usage{
load.module(name, path, quiet=FALSE)
load.pkg.module(pkgname, libname, bin.compat, quiet=FALSE)
unload.module(name, quiet=FALSE)
list.modules()
}
\arguments{
  \item{name}{name of the module to be loaded}
  \item{path}{file path to the location of the DLL. If omitted,
    the option \code{jags.moddir} is used to locate the modules}
  \item{pkgname}{name of the R package containing the module to be loaded}
  \item{libname}{path to the library directory where the R package can be found}
  \item{bin.compat}{a list generated by \code{\link{pkg.compile.flags}}
    at build time by the package containing the JAGS module to be loaded}
  \item{quiet}{a logical. If \code{TRUE}, no message will be printed
    about loading the package}
}
\examples{
list.modules()
load.module("glm")
list.modules()
unload.module("glm")
list.modules()
}
\author{Martyn Plummer}
\keyword{interface}

