\name{jags.version}
\alias{jags.version}
\alias{JAGS.version}
\title{JAGS version}
\description{
  Get the version of the JAGS library which is currently linked to this R session
}
\usage{
jags.version()
}
\value{
  The version of JAGS formatted as a package version string (see \code{\link{package_version}})
}
